requirements.txt
ipywidgets
fasta
streamlit